Differnce

+++ Sheet #1 (Data type - Conditions)/D_Difference.cpp
@@ -8,7 +8,7 @@ using namespace std;
 
 int main()
 {
-    int a, b, c, d;
+    long long int a, b, c, d;
     cin >> a >> b >> c >> d;
     cout << "Difference = " << (a * b) - (c * d);
     return 0;
